refactor : service에서는 계산만 할 수 있도록, 문자가 숫자인지, 음수인지 확인하는 부분 분리
service에서는 계산이라는 하나의 역할만 수행하도록, 검증 관련 코드는 다른 클래스로 분리
